Moved client tests up a directory
